refactor(api): extract shared request transport
