+++ src/services/apiService/common/constant.ts
@@ -1,9 +1,1 @@
-// ============= 常量定义 =============
-export const REQUEST_CONFIG = {
-  DEFAULT_PAGE_SIZE: 100,
-  MAX_PAGES: 100,
-  HEADERS: {
-    CONTENT_TYPE: "application/json",
-    PRAGMA: "no-cache",
-  },
-} as const
+export { REQUEST_CONFIG } from "~/services/apiTransport/constant"
