+++ src/services/apiTransport/errors.ts
@@ -1,0 +1,63 @@
+import { TEMP_WINDOW_HEALTH_STATUS_CODES } from "~/types"
+
+// ============= 错误处理 =============
+export const API_ERROR_CODES = {
+  HTTP_401: "HTTP_401",
+  HTTP_403: "HTTP_403",
+  HTTP_429: "HTTP_429",
+  HTTP_OTHER: "HTTP_OTHER",
+  CONTENT_TYPE_MISMATCH: "CONTENT_TYPE_MISMATCH",
+  TEMP_WINDOW_DISABLED: TEMP_WINDOW_HEALTH_STATUS_CODES.DISABLED,
+  TEMP_WINDOW_PERMISSION_REQUIRED:
+    TEMP_WINDOW_HEALTH_STATUS_CODES.PERMISSION_REQUIRED,
+  TEMP_WINDOW_WINDOWS_API_UNAVAILABLE: "TEMP_WINDOW_WINDOWS_API_UNAVAILABLE",
+  TEMP_WINDOW_WINDOW_CREATION_UNAVAILABLE:
+    "TEMP_WINDOW_WINDOW_CREATION_UNAVAILABLE",
+  TEMP_WINDOW_WINDOW_HANDLE_UNAVAILABLE:
+    "TEMP_WINDOW_WINDOW_HANDLE_UNAVAILABLE",
+  ROBOT_CHALLENGE: "ROBOT_CHALLENGE",
+  JSON_PARSE_ERROR: "JSON_PARSE_ERROR",
+  BUSINESS_ERROR: "BUSINESS_ERROR",
+  NETWORK_ERROR: "NETWORK_ERROR",
+  TOKEN_SECRET_UNAVAILABLE: "TOKEN_SECRET_UNAVAILABLE",
+  FEATURE_UNSUPPORTED: "FEATURE_UNSUPPORTED",
+  UNKNOWN: "UNKNOWN",
+} as const
+
+export type ApiErrorCode =
+  (typeof API_ERROR_CODES)[keyof typeof API_ERROR_CODES]
+
+const TEMP_WINDOW_UNSUPPORTED_ERROR_CODES = new Set<ApiErrorCode>([
+  API_ERROR_CODES.TEMP_WINDOW_WINDOWS_API_UNAVAILABLE,
+  API_ERROR_CODES.TEMP_WINDOW_WINDOW_CREATION_UNAVAILABLE,
+  API_ERROR_CODES.TEMP_WINDOW_WINDOW_HANDLE_UNAVAILABLE,
+])
+
+/**
+ * Returns whether an API error code represents a temp-context flow that
+ * required browser window creation but could not obtain it.
+ */
+export function isTempWindowUnsupportedErrorCode(
+  code?: ApiErrorCode | null,
+): code is ApiErrorCode {
+  return Boolean(code && TEMP_WINDOW_UNSUPPORTED_ERROR_CODES.has(code))
+}
+
+export class ApiError extends Error {
+  constructor(
+    message: string,
+    public statusCode?: number,
+    public endpoint?: string,
+    public code?: ApiErrorCode,
+  ) {
+    super(message)
+    this.name = "ApiError"
+  }
+
+  /**
+   * Preserves the original error code if this instance is re-labeled for UI messaging.
+   * For example, we may convert HTTP_403 into TEMP_WINDOW_DISABLED when the request
+   * could have been recovered via temp-window fallback but the feature was unavailable.
+   */
+  public originalCode?: ApiErrorCode
+}
